style: change date format to use short month names (e.g., Jan 17, 2026)

+++ src/lib/date.ts
@@ -1,5 +1,5 @@
 export function formatDate(date: Date): string {
-  return date.toLocaleDateString('en-US', { timeZone: 'UTC', year: 'numeric', month: 'long', day: 'numeric' });
+  return date.toLocaleDateString('en-US', { timeZone: 'UTC', year: 'numeric', month: 'short', day: 'numeric' });
 }
 
 export function toISODate(date: Date): string {
